fibonacci 50 times

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,18 +7,18 @@
 
 int main(void)
 {
-	long long f1, f2, fn;
+	long int f1, f2, fn;
 	int c;
 
 	f1 = 1;
 	f2 = 2;
 
-	printf("%llu, %llu", f1, f2);
+	printf("%ld, %ld", f1, f2);
 
 	for (c = 3; c <= 50; c++)
 	{
 		fn = (f1 + f2);
-		printf(", %llu", fn);
+		printf(", %ld", fn);
 
 		f1 = (f2);
 		f2 = (fn);
